requirements.txt
asgiref==3.4.1
attrs==21.2.0
backcall==0.2.0
brotlipy==0.7.0
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
charset-normalizer==2.0.4
click==8.0.1
cloudpickle==2.2.0
conllu==4.4
cryptography==3.4.7
cxxfilt==0.2.2
DAWG-Python==0.7.2
decorator==5.0.9
docopt==0.6.2
faiss-cpu==1.7.2
fastapi==0.68.0
filelock==3.0.12
fire==0.4.0
fasttext==0.9.2
Flask==2.0.1
Flask-Cors==3.0.10
h11==0.12.0
huggingface-hub==0.9.0
idna==3.2
importlib-metadata==3.10.0
iniconfig==1.1.1
itsdangerous==2.0.1
jedi==0.18.0
Jinja2==3.0.1
joblib==1.0.1
jsonschema==3.0.2
MarkupSafe==2.0.1
matplotlib-inline==0.1.2
more-itertools==8.8.0
numpy==1.21.1
olefile==0.46
packaging==21.0
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.3.1
pip==21.2.2
pluggy==0.13.1
prompt-toolkit==3.0.19
ptyprocess==0.7.0
py==1.10.0
pycosat==0.6.3
pycparser==2.20
pydantic==1.8.2
Pygments==2.9.0
pymorphy2==0.9.1
pymorphy2-dicts-ru==2.4.417127.4579844
pyOpenSSL==20.0.1
pyparsing==2.4.7
pyrsistent==0.18.0
psycopg2-binary==2.9.3
PySocks==1.7.1
pytest==6.2.4
PyYAML==5.4.1
razdel==0.5.0
regex==2021.8.3
requests==2.26.0
sacremoses==0.0.45
scikit-learn==0.24.2
SQLAlchemy==1.4.41
scipy==1.9.3
seqeval==1.2.2
six==1.16.0
starlette==0.14.2
termcolor==1.1.0
threadpoolctl==2.2.0
tokenizers==0.11.1
toml==0.10.2
torch==1.9.0
tqdm==4.62.0
traitlets==5.0.5
transformers==4.22.2
typing-extensions==3.10.0.0
urllib3==1.26.6
uvicorn==0.14.0
wcwidth==0.2.5
Werkzeug==2.0.1
wheel==0.36.2
zipp==3.5.0
pytz==2022.7.1
